apps/api/requirements.txt
fastapi[standard]
SQLAlchemy[asyncio]
asyncpg
alembic
python-dotenv
python-jose[cryptography]
bcrypt
rq
redis
pyyaml
